pyproject.toml
[project]
name = "ai-center"
version = "0.1.0"
description = "大模型AI服务中心"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "pymysql",
    "pyyaml",
    "python-dotenv",
    "pydantic",
    "pydantic-settings",
    "slowapi",
    "flasgger",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
python-version = "3.10"
